+++ app/config.py
@@ -9,6 +9,7 @@ class Settings(BaseSettings):
     templates_dir: Path = Path(__file__).resolve().parent / "templates"
     secret_key: str = Field(...)
     jwt_algorithm: str = Field(default="HS256")
+    session_duration: int = Field(default=86400)
 
     class Config:
         env_file = ".env"
