+++ app/templates/auth/login.html
@@ -1,25 +1,34 @@
 {% extends "base.html" %}
 
+
 {% block content %}
-    <h1>Login</h1>
+    <div class='row'>
+        <div class='col-md-4 text-center mx-auto'>
+            <h1>Login</h1>
 
-    {% include 'snippets/form_errors.html' %}
+            {% include 'snippets/form_errors.html' %}
 
-    {% if logged_in %}
-        <p>You are already logged in!</p>
-    {% else %}
+            {% if request.user.is_authenticated %}
+                <p>You are already logged in. Would you like to <a href='/logout'>logout</a>?</p>
+            {% else %}
+                <form method='POST'>
+                    <div>
+                        <label for='email'>Email</label>
+                        <input class='form-control' type='email' name='email' placeholder='your@email.com' />
+                    </div>
 
-        <form method='POST'>
-            <div>
-                <label for='email'>Email</label>
-                <input type='email' name='email' placeholder='your@email.com' />
-            </div>
+                    <div>
+                        <label for='password'>Password</label>
+                        <input class='form-control' type='password' name='password' placeholder='your password' />
+                    </div>
+                    <button class='btn btn-primary mt-2' type='submit' >Login </button>
 
-            <div>
-                <label for='password'>Password</label>
-                <input type='password' name='password' placeholder='your password' />
-            </div>
-            <button type='submit' >Login </button>
-        </form>
-    {% endif %}
+                </form>
+                <div class='mt-3'>
+                    <p><a href='/signup'>Do you need an account?</a></p>
+                </div>
+            {% endif %}
+        </div>
+
+    </div>
 {% endblock %}
