+++ app/templates/base.html
@@ -1,13 +1,19 @@
 <!DOCTYPE html>
 <html>
     <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
         <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous">
         <script src="https://unpkg.com/htmx.org@1.7.0/dist/htmx.min.js"></script>
         <title>Songbooks! </title>
     </head>
     <body>
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
-        {% include 'snippets/nav.html' %}
+
+        {% if request.user.is_authenticated %}
+            {% include 'snippets/nav.html' %}
+        {% endif %}
+
         {% block content %}
             <h1>Hello World</h1>
         {% endblock %}
