+++ app/templates/home.html
@@ -2,6 +2,8 @@
 
 
 {% block content %}
-
-    <h1>S</h1>
+    <h1>Welcome to the Songbook website</h1>
+    <body>
+        To experience the whole functionality please <a href='/login'>Login</a>
+    </body>
 {% endblock %}
