+++ app/utils.py
@@ -1,7 +1,12 @@
 import json
+import subprocess
+from pathlib import Path
 
 from pydantic import BaseModel
 from pydantic.error_wrappers import ValidationError
+
+from . import db
+from .songs.models import Song
 
 
 def valid_schema_data_or_error(raw_data: dict, SchemaModel: BaseModel):
@@ -19,3 +24,26 @@ def valid_schema_data_or_error(raw_data: dict, SchemaModel: BaseModel):
         except Exception:
             errors = [{"loc": "non_field_error", "msg": "Unknown error"}]
     return data, errors
+
+
+def build_song(song):
+    dest_path = Path("app/tmp/" + str(song.id))
+    pdf_path = dest_path / "source.pdf"
+    if pdf_path.exists():
+        return
+    dest_path.mkdir(parents=True, exist_ok=True)
+    source = dest_path / "source.lytex"
+
+    with source.open(mode="w") as file:
+        source_lytex = "#(ly:set-option 'crop #t)\n" + song.lytex
+        file.write(source_lytex)
+
+    subprocess.run(["lilypond", "-o", dest_path, source])
+    return
+
+
+def build_all_songs():
+    with db.get_library_session() as session:
+        songs = session.query(Song).all()
+        for song in songs:
+            build_song(song)
